+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## _Unreleased_
 - Prevent HK iframes eating damage dealt in quick succession
 - Fix bind/taunt being broken after dialogue
+- Fix not being able to open inventory on controller
+- Inherit HK controller bindings
 
 ## v0.2.0